requirements.txt
Django~=3.0
djangorestframework==3.14.0
